+++ constant/config.go
@@ -3,7 +3,7 @@ package constant
 // DefinedFieldsCount is the number of fields defined in this package.
 // You have to manually update this number when you add a new field
 // to check later if every field has a defined default value
-const DefinedFieldsCount = 47
+const DefinedFieldsCount = 48
 
 const (
 	DownloaderPath                = "downloader.path"
@@ -15,6 +15,7 @@ const (
 	DownloaderStopOnError         = "downloader.stop_on_error"
 	DownloaderDownloadCover       = "downloader.download_cover"
 	DownloaderRedownloadExisting  = "downloader.redownload_existing"
+	DownloaderReadDownloaded      = "downloader.read_downloaded"
 )
 
 const (