+++ source/page.go
@@ -29,16 +29,16 @@ type Page struct {
 	Chapter *Chapter `json:"-"`
 }
 
-func (p *Page) request() (req *http.Request, err error) {
-	req, err = http.NewRequest(http.MethodGet, p.URL, nil)
+func (p *Page) request() (*http.Request, error) {
+	req, err := http.NewRequest(http.MethodGet, p.URL, nil)
 	if err != nil {
 		log.Error(err)
-		return
+		return nil, err
 	}
 
 	req.Header.Set("Referer", p.Chapter.URL)
 	req.Header.Set("User-Agent", constant.UserAgent)
-	return
+	return req, nil
 }
 
 // Download Page contents.
@@ -108,7 +108,7 @@ func (p *Page) Close() error {
 
 // Read reads from the page contents.
 func (p *Page) Read(b []byte) (int, error) {
-	log.Debugf("Reading page contents #%d", p.Index)
+	log.Tracef("Reading page contents #%d", p.Index)
 	if p.Contents == nil {
 		err := errors.New("page not downloaded")
 		log.Error(err)